+++ src/simulation/SurfaceTemperatureSimulation.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "rendering/Scene.hpp"
+
+class SurfaceTemperatureSimulation {
+public:
+    virtual ~SurfaceTemperatureSimulation() = default;
+    virtual void setSunState(const Sun& sun) = 0;
+};