+++ src/simulation/SurfaceTemperatureStatistics.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+class SurfaceTemperatureStatistics {
+public:
+    virtual ~SurfaceTemperatureStatistics() = default;
+    [[nodiscard]] virtual float averageSurfaceTemperatureCelsius() const = 0;
+};